[스프링 - 협업 미션] 검프(김태정) 미션 제출합니다. (#33)
* feat(LinesResponse.class) : DTO 구현

* feat : 전체 distance을 구하는 메소드 구현

* refactor(PathController.class): RequestMapping 파라미터 수정

* refactor: 삭제응답 200 -> 204

* refactor: 서블릿 인터셉터 설정 및 token 검증

option, get 메소드 제외

* feat: GlobalException 설정

* feat: station 예외 로직 구현

~~역 이름은 2자 이상 20자 이하의 한글 (숫자포함. 공백허용 X) 1~~

~~역 이름 중복 예외 3~~

* feat: line 예외 로직 구현

* feat: Station 예외 처리 로직

* feat: Line 예외 처리 로직

* feat: Path 예외 처리 로직

* refactor: Name Validation 수정

* feat: 운영 디비 설정

* refactor: profiles 수정

* refactor: sql 문 수정

* refactor: db 설정

* reafctor: dataloader 수정

* refactor: DataLoder change

* feat: 라이브러리 추가

* refactor: 예외사항 추가

* feat: Detail 관련 메서드 추기

* feat(MemberService.class): 동일 이메일 검증 로직 추가

* feat(MemberTest.class): 중복된 이메일 가입 시도 예외처리

* feat: 지원하지않는 미디어 타입 예외 처리

* feat: cors 설정

* feat: cors설정 변경

* refactor: allowedOrigins 수정

* feat: 10km 미만 기본운임 금액 계산 기능 구현

* feat: 50km 초과시 추가요금 계산기능 구현

* refactor: 테스트 수정 및 매직넘버 상수화

코드 자동정렬

* refactor: extraFare 추가

* feat: 노선별 구간 추가요금 계산 구현
- sql 문법 오류 수정

* feat: 여러 노선을 거칠시 가장 높은 노선 추가금액을 적용하여 계산하는 테스트 구현

* refactor: 역,노선 이름 관련 검증로직 변경

* refactor: 모든 역 id 오름차순수정

* refactor: DB 문자열 제한 변경

* feat: 로그인한 회원의 나이에 따른 요금 계산 기능구현

* refactor: ErrorMessage Response 수정

* refactor: mysql schema

delete cascade 옵션 적용

* refactor(MemberService.class): email 로직 변경

* refactor: detail 예외 수정

* refactor: section add 리턴값 변경

* refactor: Transactional 설정

* refactor: Transactional 에러 수정

* refactor: DTO책임 도메인으로 이동

* feat: distance에 따른 금액 계산 로직 테스트 작성

* refactor(DistanceAdditionPolicy.class):  네이밍 변경

DistanceFarePolicy ->
DistanceAdditionPolicy

* feat: age에 따른 금액 계산 로직 테스트 작성

* refactor: DiscountPolicy 패키지 이동

* feat: 나이에 따른 요금 계산 로직 완료

* feat: 스크립트 파일 생성

* feat: 최단거리 경로, 캐싱 사용

* refactor: 사용하지 않는 변수삭제 및 접근지정자 변경

* refactor(LineService.class): StationService의존성 제거

* refactor: deploy.sh 파일 수정

* refactor(AgeDiscountPolicy.class): 분기문 제거 및 책임의 이동

* refactor(AgeRange.class): 원시값 포장

* refactor: graceful 옵션 설정

* refactor: 예외사항 정리 및 안쓰는 변수 삭제

* refactor: 디버깅용 logger 관련 삭제

* refactor: 의미없는 메서드화 삭제

+++ src/main/java/wooteco/subway/station/application/StationService.java
@@ -1,23 +1,38 @@
 package wooteco.subway.station.application;
 
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import wooteco.subway.exception.DuplicateException;
+import wooteco.subway.exception.InvalidInputException;
+import wooteco.subway.exception.NotFoundException;
+import wooteco.subway.line.dao.SectionDao;
 import wooteco.subway.station.dao.StationDao;
 import wooteco.subway.station.domain.Station;
 import wooteco.subway.station.dto.StationRequest;
 import wooteco.subway.station.dto.StationResponse;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 @Service
+@Transactional(readOnly = true)
 public class StationService {
-    private StationDao stationDao;
 
-    public StationService(StationDao stationDao) {
+    private final StationDao stationDao;
+    private final SectionDao sectionDao;
+
+    public StationService(StationDao stationDao, SectionDao sectionDao) {
         this.stationDao = stationDao;
+        this.sectionDao = sectionDao;
     }
 
+    @Transactional
     public StationResponse saveStation(StationRequest stationRequest) {
+        if (stationDao.existsByName(stationRequest.getName())) {
+            throw new DuplicateException("이미 존재하는 역 이름 입니다. (입력된 이름 값 : " + stationRequest.getName() + ")");
+        }
+
         Station station = stationDao.insert(stationRequest.toStation());
         return StationResponse.of(station);
     }
@@ -34,7 +49,14 @@ public class StationService {
                 .collect(Collectors.toList());
     }
 
+    @Transactional
     public void deleteStationById(Long id) {
+        if (sectionDao.existsByStationId(id)) {
+            throw new InvalidInputException("노선에 등록된 역은 삭제할 수 없습니다.");
+        }
+        if (Objects.isNull(findStationById(id))) {
+            throw new NotFoundException("삭제할 역이 등록되어 있지 않습니다.");
+        }
         stationDao.deleteById(id);
     }
 }
